% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlayout.R
\name{adjust_bb}
\alias{adjust_bb}
\title{Auxiliary function for adjusting a bounding box}
\usage{
adjust_bb(lastturn, coordslastBB, w, h)
}
\arguments{
\item{lastturn}{last turn}

\item{coordslastBB}{coordinates of the last bounding box}

\item{w}{width}

\item{h}{height}
}
\value{
Coordinates of the adjusted bounding box
}
\description{
Auxiliary function for adjusting a bounding box
}
\author{
Wayne Oldford
}
