% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2dgraphics.R
\name{arrow_2d_graphics}
\alias{arrow_2d_graphics}
\title{Arrow plot in 2d using R's base graphics}
\usage{
arrow_2d_graphics(zargs, loc = c(0.5, 0.5), angle = 60, length = 0.2,
  add = FALSE, group... = NULL, plot... = NULL, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{loc}{(x,y)-location (in (0,1)^2) of the center of the arrow}

\item{angle}{angle from the shaft to the edge of the arrow head}

\item{length}{length of the arrow in [0,1] from tip to base}

\item{add}{logical indicating whether this plot should be added to the last one}

\item{group...}{list of arguments passed to group_2d_graphics (or NULL)}

\item{plot...}{additional arguments passed to plot_region()}

\item{...}{additional arguments passed to points()}
}
\value{
invisible()
}
\description{
Arrow plot in 2d using R's base graphics
}
\seealso{
Other default 2d plot functions using R's base graphics: \code{\link{axes_2d_graphics}},
  \code{\link{density_2d_graphics}},
  \code{\link{group_2d_graphics}},
  \code{\link{label_2d_graphics}},
  \code{\link{points_2d_graphics}},
  \code{\link{qq_2d_graphics}},
  \code{\link{rect_2d_graphics}}

Other default 2d plot functions: \code{\link{arrow_2d_grid}},
  \code{\link{arrow_2d_loon}},
  \code{\link{axes_2d_graphics}},
  \code{\link{axes_2d_grid}}, \code{\link{axes_2d_loon}},
  \code{\link{density_2d_graphics}},
  \code{\link{density_2d_grid}},
  \code{\link{density_2d_loon}}, \code{\link{extract_2d}},
  \code{\link{group_2d_graphics}},
  \code{\link{group_2d_grid}}, \code{\link{group_2d_loon}},
  \code{\link{label_2d_graphics}},
  \code{\link{label_2d_grid}}, \code{\link{label_2d_loon}},
  \code{\link{points_2d_graphics}},
  \code{\link{points_2d_grid}},
  \code{\link{points_2d_loon}},
  \code{\link{qq_2d_graphics}}, \code{\link{qq_2d_grid}},
  \code{\link{rect_2d_graphics}},
  \code{\link{rect_2d_grid}}, \code{\link{rect_2d_loon}}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 2d plot functions}
\concept{default 2d plot functions using R's base graphics}
