% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{as_numeric}
\alias{as_numeric}
\title{A list of columns}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{A list of columns}
}
\value{
A list where each column is converted to data (range() works,
        can be plotted, etc.)
}
\description{
A list of columns
}
\note{
See plot.default -> xy.coords()
}
\author{
Marius Hofert
}
