% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgraphics.R
\name{boxplot_1d_graphics}
\alias{boxplot_1d_graphics}
\title{Box plot in 1d using R's base graphics}
\usage{
boxplot_1d_graphics(
  zargs,
  cex = 0.4,
  range = NULL,
  axes = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{zargs}{The argument list as passed from \code{\link{zenplot}()}}

\item{cex}{The character expansion factor}

\item{range}{A numerical value which determines how far the plot whiskers extend.
If NULL, the whiskers (range) grows with sample size.}

\item{axes}{A logicial indicating whether axes should be drawn}

\item{add}{A logical indicating whether this plot should be added to the last one}

\item{...}{Additional arguments passed to boxplot()}
}
\value{
invisible()
}
\description{
Box plot in 1d using R's base graphics
}
\seealso{
Other default 1d plot functions using R's base graphics: 
\code{\link{arrow_1d_graphics}()},
\code{\link{density_1d_graphics}()},
\code{\link{hist_1d_graphics}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{label_1d_graphics}()},
\code{\link{lines_1d_graphics}()},
\code{\link{points_1d_graphics}()},
\code{\link{rect_1d_graphics}()},
\code{\link{rug_1d_graphics}()}

Other default 1d plot functions: 
\code{\link{arrow_1d_graphics}()},
\code{\link{arrow_1d_grid}()},
\code{\link{arrow_1d_loon}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{boxplot_1d_loon}()},
\code{\link{density_1d_graphics}()},
\code{\link{density_1d_grid}()},
\code{\link{density_1d_loon}()},
\code{\link{extract_1d}()},
\code{\link{hist_1d_graphics}()},
\code{\link{hist_1d_grid}()},
\code{\link{hist_1d_loon}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{jitter_1d_grid}()},
\code{\link{jitter_1d_loon}()},
\code{\link{label_1d_graphics}()},
\code{\link{label_1d_grid}()},
\code{\link{label_1d_loon}()},
\code{\link{lines_1d_graphics}()},
\code{\link{lines_1d_grid}()},
\code{\link{lines_1d_loon}()},
\code{\link{points_1d_graphics}()},
\code{\link{points_1d_grid}()},
\code{\link{points_1d_loon}()},
\code{\link{rect_1d_graphics}()},
\code{\link{rect_1d_grid}()},
\code{\link{rect_1d_loon}()},
\code{\link{rug_1d_graphics}()},
\code{\link{rug_1d_grid}()},
\code{\link{rug_1d_loon}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using R's base graphics}
