% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_zargs}
\alias{check_zargs}
\title{Checking whether certain arguments appear in zargs}
\usage{
check_zargs(zargs, ...)
}
\arguments{
\item{zargs}{The argument list as passed from zenplot()}

\item{...}{The arguments to be checked for presence in zargs}
}
\value{
A logical indicating whether some arguments are missing in zargs
}
\description{
Checking whether certain arguments appear in zargs
}
\seealso{
Other tools for constructing your own plot1d and plot2d functions: 
\code{\link{burst_aux}()},
\code{\link{burst}()},
\code{\link{extract_1d}()},
\code{\link{extract_2d}()},
\code{\link{plot_indices}()}
}
\author{
Marius Hofert
}
\concept{tools for constructing your own plot1d and plot2d functions}
