% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgrid.R
\name{density_1d_grid}
\alias{density_1d_grid}
\title{Density plot in 1d using the grid package}
\usage{
density_1d_grid(zargs, density... = NULL, offset = 0.08,
  draw = FALSE, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{density...}{list of arguments for density()}

\item{offset}{numerical value in \deqn{[0, 0.5]} used to offset
the density within the height 1 box in which it appears}

\item{draw}{logical indicating whether drawing should take place}

\item{...}{additional arguments passed to gpar()}
}
\value{
grob (invisibly)
}
\description{
Density plot in 1d using the grid package
}
\seealso{
Other default 1d plot functions using the grid package: \code{\link{arrow_1d_grid}},
  \code{\link{boxplot_1d_grid}},
  \code{\link{hist_1d_grid}}, \code{\link{jitter_1d_grid}},
  \code{\link{label_1d_grid}}, \code{\link{lines_1d_grid}},
  \code{\link{points_1d_grid}}, \code{\link{rect_1d_grid}},
  \code{\link{rug_1d_grid}}

Other default 1d plot functions: \code{\link{arrow_1d_graphics}},
  \code{\link{arrow_1d_grid}}, \code{\link{arrow_1d_loon}},
  \code{\link{boxplot_1d_graphics}},
  \code{\link{boxplot_1d_grid}},
  \code{\link{boxplot_1d_loon}},
  \code{\link{density_1d_graphics}},
  \code{\link{density_1d_loon}}, \code{\link{extract_1d}},
  \code{\link{hist_1d_graphics}},
  \code{\link{hist_1d_grid}}, \code{\link{hist_1d_loon}},
  \code{\link{jitter_1d_graphics}},
  \code{\link{jitter_1d_grid}},
  \code{\link{jitter_1d_loon}},
  \code{\link{label_1d_graphics}},
  \code{\link{label_1d_grid}}, \code{\link{label_1d_loon}},
  \code{\link{lines_1d_graphics}},
  \code{\link{lines_1d_grid}}, \code{\link{lines_1d_loon}},
  \code{\link{points_1d_graphics}},
  \code{\link{points_1d_grid}},
  \code{\link{points_1d_loon}},
  \code{\link{rect_1d_graphics}},
  \code{\link{rect_1d_grid}}, \code{\link{rect_1d_loon}},
  \code{\link{rug_1d_graphics}}, \code{\link{rug_1d_grid}},
  \code{\link{rug_1d_loon}}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using the grid package}
