% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_1d}
\alias{extract_1d}
\title{Extracting information for our default/provided plot1d()}
\usage{
extract_1d(zargs)
}
\arguments{
\item{zargs}{The argument list as passed from zenplot(); this must at least
contain 'x', 'orientations', 'vars', 'num', 'lim' and 'labs'.}
}
\value{
A list with
        - the data x to plot in the 1d plot;
        - a list with all columns of x;
        - the group numbers (for each column of x);
        - the variable numbers (for each column of x);
        - the group labels (for each column of x);
        - the variable labels (for each column of x);
        - horizontal (logical);
        - the (x-)axis limits
}
\description{
Extracting information for our default/provided plot1d()
}
\note{
Performance critical
}
\seealso{
Other tools for constructing your own plot1d and plot2d functions: \code{\link{burst_aux}},
  \code{\link{burst}}, \code{\link{check_zargs}},
  \code{\link{extract_2d}}, \code{\link{plot_indices}}
}
\author{
Marius Hofert
}
\concept{tools for constructing your own plot1d and plot2d functions}
