% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpath.R
\name{get_path}
\alias{get_path}
\title{Computing the path according to the provided method}
\usage{
get_path(turns = NULL, n2dcols = c("letter", "square", "A4", "golden",
  "legal"), n2dplots, method = c("tidy", "double.zigzag",
  "single.zigzag", "rectangular"), first1d = TRUE, last1d = TRUE)
}
\arguments{
\item{turns}{The turns}

\item{n2dcols}{The number of columns of 2d plots (>= 1) or one of "letter", "square",
"A4", "golden", "legal". Note that n2dcols is ignored if turns is not NULL.}

\item{n2dplots}{The number of 2d plots to be laid out}

\item{method}{A character string indicating the method according to which the
path is built}

\item{first1d}{A logical indicating whether the first 1d plot should be plotted}

\item{last1d}{A logical indicating whether the last 1d plot should be plotted}
}
\value{
the path, a list containing the turns, the positions (indices in the
        occupancy matrix) and the the occupancy matrix
}
\description{
Computing the path according to the provided method
}
\author{
Marius Hofert and Wayne Oldford
}
