% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpath.R
\name{get_zigzag_turns}
\alias{get_zigzag_turns}
\title{Compute turns for zigzag}
\usage{
get_zigzag_turns(nPlots, n2dcols, method = c("tidy", "double.zigzag",
  "single.zigzag"))
}
\arguments{
\item{nPlots}{total number of plots}

\item{n2dcols}{number of columns of 2d plots (>= 1)}

\item{method}{character string indicating which zigzag method to use}
}
\value{
turns
}
\description{
Compute turns for zigzag
}
\author{
Marius Hofert and Wayne Oldford
}
