% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgraphics.R
\name{hist_1d_graphics}
\alias{hist_1d_graphics}
\title{Histogram as 1d plot using R's base graphics}
\usage{
hist_1d_graphics(zargs, breaks = NULL, length.out = 21, col = NULL,
  axes = FALSE, add = TRUE, plot... = NULL, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{breaks}{see ?hist; the default is 20 equi-width bins covering the data range}

\item{length.out}{number of break points if breaks = NULL}

\item{col}{vector of colors for the bars or bar components; see ?barplot}

\item{axes}{logicial indicating whether axes should be drawn}

\item{add}{logical indicating whether this plot should be added to the last one}

\item{plot...}{additional arguments passed to plot_region()}

\item{...}{additional arguments passed to barplot()}
}
\value{
invisible()
}
\description{
Histogram as 1d plot using R's base graphics
}
\seealso{
Other default 1d plot functions using R's base graphics: \code{\link{arrow_1d_graphics}},
  \code{\link{boxplot_1d_graphics}},
  \code{\link{density_1d_graphics}},
  \code{\link{jitter_1d_graphics}},
  \code{\link{label_1d_graphics}},
  \code{\link{lines_1d_graphics}},
  \code{\link{points_1d_graphics}},
  \code{\link{rect_1d_graphics}},
  \code{\link{rug_1d_graphics}}

Other default 1d plot functions: \code{\link{arrow_1d_graphics}},
  \code{\link{arrow_1d_grid}}, \code{\link{arrow_1d_loon}},
  \code{\link{boxplot_1d_graphics}},
  \code{\link{boxplot_1d_grid}},
  \code{\link{boxplot_1d_loon}},
  \code{\link{density_1d_graphics}},
  \code{\link{density_1d_grid}},
  \code{\link{density_1d_loon}}, \code{\link{extract_1d}},
  \code{\link{hist_1d_grid}}, \code{\link{hist_1d_loon}},
  \code{\link{jitter_1d_graphics}},
  \code{\link{jitter_1d_grid}},
  \code{\link{jitter_1d_loon}},
  \code{\link{label_1d_graphics}},
  \code{\link{label_1d_grid}}, \code{\link{label_1d_loon}},
  \code{\link{lines_1d_graphics}},
  \code{\link{lines_1d_grid}}, \code{\link{lines_1d_loon}},
  \code{\link{points_1d_graphics}},
  \code{\link{points_1d_grid}},
  \code{\link{points_1d_loon}},
  \code{\link{rect_1d_graphics}},
  \code{\link{rect_1d_grid}}, \code{\link{rect_1d_loon}},
  \code{\link{rug_1d_graphics}}, \code{\link{rug_1d_grid}},
  \code{\link{rug_1d_loon}}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using R's base graphics}
