% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgrid.R
\name{hist_1d_grid}
\alias{hist_1d_grid}
\title{Histogram in 1d using the grid package}
\usage{
hist_1d_grid(
  zargs,
  breaks = NULL,
  length.out = 21,
  col = NULL,
  fill = NULL,
  draw = FALSE,
  ...
)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{breaks}{see ?hist; the default is 20 equi-width bins covering the data range}

\item{length.out}{number of break points if breaks = NULL}

\item{col}{colour of the histogram bar interiors, unless fill is specified, then
this is the colour of the border}

\item{fill}{logical passed to the underlying rectGrob()}

\item{draw}{logical indicating whether drawing should take place}

\item{...}{additional arguments passed to gpar()}
}
\value{
grob (invisibly)
}
\description{
Histogram in 1d using the grid package
}
\seealso{
Other default 1d plot functions using the grid package: 
\code{\link{arrow_1d_grid}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{density_1d_grid}()},
\code{\link{jitter_1d_grid}()},
\code{\link{label_1d_grid}()},
\code{\link{lines_1d_grid}()},
\code{\link{points_1d_grid}()},
\code{\link{rect_1d_grid}()},
\code{\link{rug_1d_grid}()}

Other default 1d plot functions: 
\code{\link{arrow_1d_graphics}()},
\code{\link{arrow_1d_grid}()},
\code{\link{arrow_1d_loon}()},
\code{\link{boxplot_1d_graphics}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{boxplot_1d_loon}()},
\code{\link{density_1d_graphics}()},
\code{\link{density_1d_grid}()},
\code{\link{density_1d_loon}()},
\code{\link{extract_1d}()},
\code{\link{hist_1d_graphics}()},
\code{\link{hist_1d_loon}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{jitter_1d_grid}()},
\code{\link{jitter_1d_loon}()},
\code{\link{label_1d_graphics}()},
\code{\link{label_1d_grid}()},
\code{\link{label_1d_loon}()},
\code{\link{lines_1d_graphics}()},
\code{\link{lines_1d_grid}()},
\code{\link{lines_1d_loon}()},
\code{\link{points_1d_graphics}()},
\code{\link{points_1d_grid}()},
\code{\link{points_1d_loon}()},
\code{\link{rect_1d_graphics}()},
\code{\link{rect_1d_grid}()},
\code{\link{rect_1d_loon}()},
\code{\link{rug_1d_graphics}()},
\code{\link{rug_1d_grid}()},
\code{\link{rug_1d_loon}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using the grid package}
