% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is.standard}
\alias{is.standard}
\title{Check Argument for Being a Vector, Matrix, Data Frame or a List of such}
\usage{
is.standard(x)
}
\arguments{
\item{x}{A vector, matrix, data.frame or list of such}
}
\value{
A logical indicating whether x is of the above type
}
\description{
Check Argument for Being a Vector, Matrix, Data Frame or a List of such
}
\seealso{
Other zenplot technical tools: \code{\link{convert_occupancy}},
  \code{\link{n2dcols_aux}}, \code{\link{num_cols}},
  \code{\link{turn_checker}}
}
\author{
Marius Hofert
}
\concept{zenplot technical tools}
