% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{l_ispace_config}
\alias{l_ispace_config}
\title{Configuring a loon plot to accommodate ispace}
\usage{
l_ispace_config(baseplot, ispace = NULL, x = NULL, y = NULL,
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{baseplot}{The plot to be modified}

\item{ispace}{The inner space (in [0,1])}

\item{x}{The x data}

\item{y}{The y data}

\item{xlim}{The x-axis limits; if NULL, the data limits are used}

\item{ylim}{The y-axis limits; if NULL, the data limits are used}

\item{...}{Additional arguments passed to loon::l_configure}
}
\value{
The baseplot
}
\description{
Configuring a loon plot to accommodate ispace
}
\seealso{
Other graphical tools: \code{\link{na_omit_loon}},
  \code{\link{plot_region}}, \code{\link{vport}},
  \code{\link{zenarrow}}
}
\author{
R. W. Oldford
}
\concept{graphical tools}
