% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgrid.R
\name{label_1d_grid}
\alias{label_1d_grid}
\title{Label plot in 1d using the grid package}
\usage{
label_1d_grid(
  zargs,
  loc = c(0.5, 0.5),
  label = NULL,
  cex = 0.66,
  box = FALSE,
  box.width = 1,
  box.height = 1,
  draw = FALSE,
  ...
)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{loc}{(x,y)-location in [0,1]^2; 0 corresponds to left, 1 to right (in
the direction of the path)}

\item{label}{label to be used}

\item{cex}{character expansion factor}

\item{box}{logical indicating whether a box should be drawn around
the text}

\item{box.width}{width of the box}

\item{box.height}{height of the box}

\item{draw}{logical indicating whether drawing should take place}

\item{...}{additional arguments passed to gpar()}
}
\value{
grob (invisibly)
}
\description{
Label plot in 1d using the grid package
}
\seealso{
Other default 1d plot functions using the grid package: 
\code{\link{arrow_1d_grid}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{density_1d_grid}()},
\code{\link{hist_1d_grid}()},
\code{\link{jitter_1d_grid}()},
\code{\link{lines_1d_grid}()},
\code{\link{points_1d_grid}()},
\code{\link{rect_1d_grid}()},
\code{\link{rug_1d_grid}()}

Other default 1d plot functions: 
\code{\link{arrow_1d_graphics}()},
\code{\link{arrow_1d_grid}()},
\code{\link{arrow_1d_loon}()},
\code{\link{boxplot_1d_graphics}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{boxplot_1d_loon}()},
\code{\link{density_1d_graphics}()},
\code{\link{density_1d_grid}()},
\code{\link{density_1d_loon}()},
\code{\link{extract_1d}()},
\code{\link{hist_1d_graphics}()},
\code{\link{hist_1d_grid}()},
\code{\link{hist_1d_loon}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{jitter_1d_grid}()},
\code{\link{jitter_1d_loon}()},
\code{\link{label_1d_graphics}()},
\code{\link{label_1d_loon}()},
\code{\link{lines_1d_graphics}()},
\code{\link{lines_1d_grid}()},
\code{\link{lines_1d_loon}()},
\code{\link{points_1d_graphics}()},
\code{\link{points_1d_grid}()},
\code{\link{points_1d_loon}()},
\code{\link{rect_1d_graphics}()},
\code{\link{rect_1d_grid}()},
\code{\link{rect_1d_loon}()},
\code{\link{rug_1d_graphics}()},
\code{\link{rug_1d_grid}()},
\code{\link{rug_1d_loon}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using the grid package}
