% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2dgrid.R
\name{label_2d_grid}
\alias{label_2d_grid}
\title{Label plot in 2d using the grid package}
\usage{
label_2d_grid(
  zargs,
  loc = c(0.98, 0.05),
  label = NULL,
  cex = 0.66,
  just = c("right", "bottom"),
  rot = 0,
  box = FALSE,
  box.width = 1,
  box.height = 1,
  group... = list(cex = cex),
  draw = FALSE,
  ...
)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{loc}{(x,y)-location in [0,1]^2; 0 corresponds to left, 1 to right (in
the direction of the path)}

\item{label}{label to be used}

\item{cex}{character expansion factor}

\item{just}{(x,y)-justification of the label}

\item{rot}{rotation of the label}

\item{box}{logical indicating whether a box should be drawn}

\item{box.width}{width of the box}

\item{box.height}{height of the box}

\item{group...}{list of arguments passed to group_2d_grid (or NULL)}

\item{draw}{logical indicating whether drawing should take place}

\item{...}{additional arguments passed to gpar()}
}
\value{
grob (invisibly)
}
\description{
Label plot in 2d using the grid package
}
\seealso{
Other default 2d plot functions using the grid package: 
\code{\link{arrow_2d_grid}()},
\code{\link{axes_2d_grid}()},
\code{\link{density_2d_grid}()},
\code{\link{group_2d_grid}()},
\code{\link{points_2d_grid}()},
\code{\link{qq_2d_grid}()},
\code{\link{rect_2d_grid}()}

Other default 2d plot functions: 
\code{\link{arrow_2d_graphics}()},
\code{\link{arrow_2d_grid}()},
\code{\link{arrow_2d_loon}()},
\code{\link{axes_2d_graphics}()},
\code{\link{axes_2d_grid}()},
\code{\link{axes_2d_loon}()},
\code{\link{density_2d_graphics}()},
\code{\link{density_2d_grid}()},
\code{\link{density_2d_loon}()},
\code{\link{extract_2d}()},
\code{\link{group_2d_graphics}()},
\code{\link{group_2d_grid}()},
\code{\link{group_2d_loon}()},
\code{\link{label_2d_graphics}()},
\code{\link{label_2d_loon}()},
\code{\link{points_2d_graphics}()},
\code{\link{points_2d_grid}()},
\code{\link{points_2d_loon}()},
\code{\link{qq_2d_graphics}()},
\code{\link{qq_2d_grid}()},
\code{\link{rect_2d_graphics}()},
\code{\link{rect_2d_grid}()},
\code{\link{rect_2d_loon}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 2d plot functions}
\concept{default 2d plot functions using the grid package}
