% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgraphics.R
\name{layout_1d_graphics}
\alias{layout_1d_graphics}
\title{Layout plot in 1d}
\usage{
layout_1d_graphics(zargs, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{...}{additional arguments passed to label_1d_graphics()}
}
\value{
invisible()
}
\description{
Layout plot in 1d
}
\author{
Marius Hofert and Wayne Oldford
}
