% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgrid.R
\name{layout_1d_grid}
\alias{layout_1d_grid}
\title{Layout plot in 1d using the grid package}
\usage{
layout_1d_grid(zargs, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{...}{additional arguments passed to label_1d_grid()}
}
\value{
grob (invisibly)
}
\description{
Layout plot in 1d using the grid package
}
\author{
Marius Hofert and Wayne Oldford
}
