% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dloon.R
\name{layout_1d_loon}
\alias{layout_1d_loon}
\title{Layout plot in 1d using the interactive loon package}
\usage{
layout_1d_loon(zargs, ...)
}
\arguments{
\item{zargs}{The argument list as passed from \code{\link{zenplot}()}}

\item{...}{Additional arguments passed to label_1d_loon()}
}
\value{
invisible()
}
\description{
Layout plot in 1d using the interactive loon package
}
\author{
Marius Hofert and Wayne Oldford
}
