% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2dgraphics.R
\name{layout_2d_graphics}
\alias{layout_2d_graphics}
\title{Layout plot in 2d}
\usage{
layout_2d_graphics(zargs, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{...}{additional arguments passed to label_2d_graphics()}
}
\value{
invisible()
}
\description{
Layout plot in 2d
}
\note{
Here we also pass '...' to group_2d_grid() (to easily adjust
      font size etc.)
}
\author{
Marius Hofert and Wayne Oldford
}
