% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2dgrid.R
\name{layout_2d_grid}
\alias{layout_2d_grid}
\title{Layout plot in 2d using the grid package}
\usage{
layout_2d_grid(zargs, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{...}{additional arguments passed to label_2d_grid()}
}
\value{
grob (invisibly)
}
\description{
Layout plot in 2d using the grid package
}
\note{
Here we also pass '...' to group_2d_grid() (to easily adjust
      font size etc.)
}
\author{
Marius Hofert and Wayne Oldford
}
