% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2dloon.R
\name{layout_2d_loon}
\alias{layout_2d_loon}
\title{Layout plot in 2d using the interactive loon package}
\usage{
layout_2d_loon(zargs, ...)
}
\arguments{
\item{zargs}{The argument list as passed from \code{\link{zenplot}()}}

\item{...}{Additional arguments passed to label_2d_grid()}
}
\value{
A loon plot
}
\description{
Layout plot in 2d using the interactive loon package
}
\note{
Here we also pass '...' to group_2d_loon() (to easily adjust
      font size etc.)
}
\author{
Marius Hofert and Wayne Oldford
}
