% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{n2dcols_aux}
\alias{n2dcols_aux}
\title{Auxiliary Function for Constructing Default n2dcols}
\usage{
n2dcols_aux(n2dplots, method = c("letter", "square", "A4", "golden",
  "legal"))
}
\arguments{
\item{n2dplots}{The number of variates (= nfaces)}

\item{method}{One of "letter", "square", "A4", "golden", "legal"}
}
\value{
An odd integer for n2dcols
}
\description{
Auxiliary Function for Constructing Default n2dcols
}
\seealso{
Other zenplot technical tools: \code{\link{convert_occupancy}},
  \code{\link{is.standard}}, \code{\link{num_cols}},
  \code{\link{turn_checker}}
}
\author{
Wayne Oldford
}
\concept{zenplot technical tools}
