% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpath.R
\name{next_move_tidy}
\alias{next_move_tidy}
\title{Determine the next position to move to and the turn out of there}
\usage{
next_move_tidy(plotNo, nPlots, curpath)
}
\arguments{
\item{plotNo}{current plot number}

\item{nPlots}{total number of plots}

\item{curpath}{the current path}
}
\value{
a list containing the next position to move to (nextpos) and the turn
        out of there (nextout); Interpretation:
        nextpos: position of plot number plotNo+1 in the (non-trimmed) occupancy matrix
        nextout: turn out of nextpos
}
\description{
Determine the next position to move to and the turn out of there
}
\note{
- This assumes that the last plot is a 1d plot!
      - It also assumes that first1d = TRUE; will be adapted later in get_path()
        in case first1d = FALSE.
      - We start in (1, 2) and also have an additional last column in the occupancy
        matrix to have the first and last column left in case we end up there with
        the last 1d plot; this cannot happen for 'zigzag' but for 'tidy'.
}
\author{
Marius Hofert and Wayne Oldford
}
