% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{num_cols}
\alias{num_cols}
\title{Determine the number of columns if is.standard(x)}
\usage{
num_cols(x)
}
\arguments{
\item{x}{A numeric vector, matrix, data.frame or a list of such.}
}
\value{
The number of data columns of 'x'
}
\description{
Determine the number of columns if is.standard(x)
}
\seealso{
Other zenplot technical tools: \code{\link{convert_occupancy}},
  \code{\link{is.standard}}, \code{\link{n2dcols_aux}},
  \code{\link{turn_checker}}
}
\author{
Marius Hofert
}
\concept{zenplot technical tools}
