% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_exists}
\alias{plot_exists}
\title{Check whether functions (plot*d to zenplot()) exist}
\usage{
plot_exists(x)
}
\arguments{
\item{x}{arguments plot1d or plot2d of zenplot()}
}
\value{
logical indicating whether x exists
}
\description{
Check whether functions (plot*d to zenplot()) exist
}
\note{
Check first whether it's a function (have to rely on it being able to be evaluated,
      cannot do more checks then) or, if a string, whether it exists
}
\author{
Marius Hofert
}
