% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_region}
\alias{plot_region}
\title{Function to set up the plot region for graphics plots}
\usage{
plot_region(xlim, ylim, plot... = NULL)
}
\arguments{
\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{plot...}{arguments passed to the underlying \code{\link{plot}()}}
}
\value{
\code{\link{invisible}()}
}
\description{
Auxiliary function for setting up the plot region
of 1d and 2d graphics plots.
}
\details{
This is an auxiliary function used by the
provided \pkg{graphics}-related 1d and 2d plots.
}
\seealso{
Other graphical tools: 
\code{\link{l_ispace_config}()},
\code{\link{na_omit_loon}()},
\code{\link{vport}()},
\code{\link{zenarrow}()}
}
\author{
Marius Hofert
}
\concept{graphical tools}
\keyword{dplot}
