% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dloon.R
\name{points_1d_loon}
\alias{points_1d_loon}
\title{Dot plot in 1d using the interactive loon package}
\usage{
points_1d_loon(zargs, linkingGroup = NULL, linkingKey = NULL,
  showLabels = FALSE, showScales = FALSE, showGuides = FALSE,
  glyph = "ocircle", itemLabel = NULL, showItemLabels = TRUE,
  parent = NULL, ...)
}
\arguments{
\item{zargs}{The argument list as passed from \code{\link{zenplot}()}}

\item{linkingGroup}{A string specifying the initial group of plots to be
linked to this plot}

\item{linkingKey}{List of IDs to link on}

\item{showLabels}{Logical determining whether axis labels are displayed}

\item{showScales}{Logical determining whether scales are displayed}

\item{showGuides}{Logical determining whether the background guidelines are
displayed}

\item{glyph}{The plot glyph}

\item{itemLabel}{A vector of strings to serve as the item labels}

\item{showItemLabels}{Logical determing whether item labels display on mouse
hover}

\item{parent}{The tk parent for this loon plot widget}

\item{...}{Additional parameters passed to loon::l_plot()}
}
\value{
A loon loon::l_plot(...)
}
\description{
Dot plot in 1d using the interactive loon package
}
\seealso{
Other default 1d plot functions using the interactive loon package: \code{\link{arrow_1d_loon}},
  \code{\link{boxplot_1d_loon}},
  \code{\link{density_1d_loon}},
  \code{\link{hist_1d_loon}}, \code{\link{jitter_1d_loon}},
  \code{\link{label_1d_loon}}, \code{\link{lines_1d_loon}},
  \code{\link{rect_1d_loon}}, \code{\link{rug_1d_loon}}

Other default 1d plot functions: \code{\link{arrow_1d_graphics}},
  \code{\link{arrow_1d_grid}}, \code{\link{arrow_1d_loon}},
  \code{\link{boxplot_1d_graphics}},
  \code{\link{boxplot_1d_grid}},
  \code{\link{boxplot_1d_loon}},
  \code{\link{density_1d_graphics}},
  \code{\link{density_1d_grid}},
  \code{\link{density_1d_loon}}, \code{\link{extract_1d}},
  \code{\link{hist_1d_graphics}},
  \code{\link{hist_1d_grid}}, \code{\link{hist_1d_loon}},
  \code{\link{jitter_1d_graphics}},
  \code{\link{jitter_1d_grid}},
  \code{\link{jitter_1d_loon}},
  \code{\link{label_1d_graphics}},
  \code{\link{label_1d_grid}}, \code{\link{label_1d_loon}},
  \code{\link{lines_1d_graphics}},
  \code{\link{lines_1d_grid}}, \code{\link{lines_1d_loon}},
  \code{\link{points_1d_graphics}},
  \code{\link{points_1d_grid}},
  \code{\link{rect_1d_graphics}},
  \code{\link{rect_1d_grid}}, \code{\link{rect_1d_loon}},
  \code{\link{rug_1d_graphics}}, \code{\link{rug_1d_grid}},
  \code{\link{rug_1d_loon}}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using the interactive loon package}
