% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2dgraphics.R
\name{qq_2d_graphics}
\alias{qq_2d_graphics}
\title{Quantile-quantile plot in 2d using R's base graphics}
\usage{
qq_2d_graphics(
  zargs,
  do.line = TRUE,
  lines... = NULL,
  cex = 0.4,
  box = FALSE,
  add = FALSE,
  group... = NULL,
  plot... = NULL,
  ...
)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{do.line}{logical indicating whether a line is drawn (through both
empirical c(0.25, 0.75)-quantiles)}

\item{lines...}{additional arguments passed to lines()}

\item{cex}{character expansion factor}

\item{box}{logical indicating whether a box should be drawn}

\item{add}{logical indicating whether this plot should be added to the last one}

\item{group...}{list of arguments passed to group_2d_graphics (or NULL)}

\item{plot...}{additional arguments passed to plot_region()}

\item{...}{additional arguments passed to qqplot()}
}
\value{
invisible()
}
\description{
Quantile-quantile plot in 2d using R's base graphics
}
\note{
line iff both margins are of the same *type*
}
\seealso{
Other default 2d plot functions using R's base graphics: 
\code{\link{arrow_2d_graphics}()},
\code{\link{axes_2d_graphics}()},
\code{\link{density_2d_graphics}()},
\code{\link{group_2d_graphics}()},
\code{\link{label_2d_graphics}()},
\code{\link{points_2d_graphics}()},
\code{\link{rect_2d_graphics}()}

Other default 2d plot functions: 
\code{\link{arrow_2d_graphics}()},
\code{\link{arrow_2d_grid}()},
\code{\link{arrow_2d_loon}()},
\code{\link{axes_2d_graphics}()},
\code{\link{axes_2d_grid}()},
\code{\link{axes_2d_loon}()},
\code{\link{density_2d_graphics}()},
\code{\link{density_2d_grid}()},
\code{\link{density_2d_loon}()},
\code{\link{extract_2d}()},
\code{\link{group_2d_graphics}()},
\code{\link{group_2d_grid}()},
\code{\link{group_2d_loon}()},
\code{\link{label_2d_graphics}()},
\code{\link{label_2d_grid}()},
\code{\link{label_2d_loon}()},
\code{\link{points_2d_graphics}()},
\code{\link{points_2d_grid}()},
\code{\link{points_2d_loon}()},
\code{\link{qq_2d_grid}()},
\code{\link{rect_2d_graphics}()},
\code{\link{rect_2d_grid}()},
\code{\link{rect_2d_loon}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 2d plot functions}
\concept{default 2d plot functions using R's base graphics}
