% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgraphics.R
\name{rect_1d_graphics}
\alias{rect_1d_graphics}
\title{Rectangle plot in 1d using R's base graphics}
\usage{
rect_1d_graphics(
  zargs,
  loc = c(0.5, 0.5),
  width = 1,
  height = 1,
  add = FALSE,
  plot... = NULL,
  ...
)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{loc}{(x,y)-location in [0,1]^2; 0 corresponds to left, 1 to right (in
the direction of the path)}

\item{width}{width of the rectangle (when viewed in walking direction)}

\item{height}{height of the rectangle (when viewed in walking direction)}

\item{add}{logical indicating whether this plot should be added to the last one}

\item{plot...}{additional arguments passed to plot_region()}

\item{...}{additional arguments passed to lines()}
}
\value{
invisible()
}
\description{
Rectangle plot in 1d using R's base graphics
}
\seealso{
Other default 1d plot functions using R's base graphics: 
\code{\link{arrow_1d_graphics}()},
\code{\link{boxplot_1d_graphics}()},
\code{\link{density_1d_graphics}()},
\code{\link{hist_1d_graphics}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{label_1d_graphics}()},
\code{\link{lines_1d_graphics}()},
\code{\link{points_1d_graphics}()},
\code{\link{rug_1d_graphics}()}

Other default 1d plot functions: 
\code{\link{arrow_1d_graphics}()},
\code{\link{arrow_1d_grid}()},
\code{\link{arrow_1d_loon}()},
\code{\link{boxplot_1d_graphics}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{boxplot_1d_loon}()},
\code{\link{density_1d_graphics}()},
\code{\link{density_1d_grid}()},
\code{\link{density_1d_loon}()},
\code{\link{extract_1d}()},
\code{\link{hist_1d_graphics}()},
\code{\link{hist_1d_grid}()},
\code{\link{hist_1d_loon}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{jitter_1d_grid}()},
\code{\link{jitter_1d_loon}()},
\code{\link{label_1d_graphics}()},
\code{\link{label_1d_grid}()},
\code{\link{label_1d_loon}()},
\code{\link{lines_1d_graphics}()},
\code{\link{lines_1d_grid}()},
\code{\link{lines_1d_loon}()},
\code{\link{points_1d_graphics}()},
\code{\link{points_1d_grid}()},
\code{\link{points_1d_loon}()},
\code{\link{rect_1d_grid}()},
\code{\link{rect_1d_loon}()},
\code{\link{rug_1d_graphics}()},
\code{\link{rug_1d_grid}()},
\code{\link{rug_1d_loon}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using R's base graphics}
