% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dgrid.R
\name{rect_1d_grid}
\alias{rect_1d_grid}
\title{Rectangle plot in 1d using the grid package}
\usage{
rect_1d_grid(zargs, loc = c(0.5, 0.5), width = 1, height = 1,
  draw = FALSE, ...)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}}

\item{loc}{(x,y)-location of the rectangle}

\item{width}{width of the rectangle (when viewed in walking direction)}

\item{height}{height of the rectangle (when viewed in walking direction)}

\item{draw}{logical indicating whether drawing should take place}

\item{...}{additional arguments passed to gpar()}
}
\value{
grob (invisibly)
}
\description{
Rectangle plot in 1d using the grid package
}
\seealso{
Other default 1d plot functions using the grid package: \code{\link{arrow_1d_grid}},
  \code{\link{boxplot_1d_grid}},
  \code{\link{density_1d_grid}},
  \code{\link{hist_1d_grid}}, \code{\link{jitter_1d_grid}},
  \code{\link{label_1d_grid}}, \code{\link{lines_1d_grid}},
  \code{\link{points_1d_grid}}, \code{\link{rug_1d_grid}}

Other default 1d plot functions: \code{\link{arrow_1d_graphics}},
  \code{\link{arrow_1d_grid}}, \code{\link{arrow_1d_loon}},
  \code{\link{boxplot_1d_graphics}},
  \code{\link{boxplot_1d_grid}},
  \code{\link{boxplot_1d_loon}},
  \code{\link{density_1d_graphics}},
  \code{\link{density_1d_grid}},
  \code{\link{density_1d_loon}}, \code{\link{extract_1d}},
  \code{\link{hist_1d_graphics}},
  \code{\link{hist_1d_grid}}, \code{\link{hist_1d_loon}},
  \code{\link{jitter_1d_graphics}},
  \code{\link{jitter_1d_grid}},
  \code{\link{jitter_1d_loon}},
  \code{\link{label_1d_graphics}},
  \code{\link{label_1d_grid}}, \code{\link{label_1d_loon}},
  \code{\link{lines_1d_graphics}},
  \code{\link{lines_1d_grid}}, \code{\link{lines_1d_loon}},
  \code{\link{points_1d_graphics}},
  \code{\link{points_1d_grid}},
  \code{\link{points_1d_loon}},
  \code{\link{rect_1d_graphics}},
  \code{\link{rect_1d_loon}},
  \code{\link{rug_1d_graphics}}, \code{\link{rug_1d_grid}},
  \code{\link{rug_1d_loon}}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using the grid package}
