% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1dloon.R
\name{rug_1d_loon}
\alias{rug_1d_loon}
\title{Rug plot in 1d using the interactive loon package}
\usage{
rug_1d_loon(zargs, ...)
}
\arguments{
\item{zargs}{The argument list as passed from \code{\link{zenplot}()}}

\item{...}{Additional parameters passed to loon::l_plot()}
}
\value{
A loon loon::l_plot(...)
}
\description{
Rug plot in 1d using the interactive loon package
}
\note{
Just calls points_1d_loon with glyph = "osquare" to preserve linking
}
\seealso{
Other default 1d plot functions using the interactive loon package: 
\code{\link{arrow_1d_loon}()},
\code{\link{boxplot_1d_loon}()},
\code{\link{density_1d_loon}()},
\code{\link{hist_1d_loon}()},
\code{\link{jitter_1d_loon}()},
\code{\link{label_1d_loon}()},
\code{\link{lines_1d_loon}()},
\code{\link{points_1d_loon}()},
\code{\link{rect_1d_loon}()}

Other default 1d plot functions: 
\code{\link{arrow_1d_graphics}()},
\code{\link{arrow_1d_grid}()},
\code{\link{arrow_1d_loon}()},
\code{\link{boxplot_1d_graphics}()},
\code{\link{boxplot_1d_grid}()},
\code{\link{boxplot_1d_loon}()},
\code{\link{density_1d_graphics}()},
\code{\link{density_1d_grid}()},
\code{\link{density_1d_loon}()},
\code{\link{extract_1d}()},
\code{\link{hist_1d_graphics}()},
\code{\link{hist_1d_grid}()},
\code{\link{hist_1d_loon}()},
\code{\link{jitter_1d_graphics}()},
\code{\link{jitter_1d_grid}()},
\code{\link{jitter_1d_loon}()},
\code{\link{label_1d_graphics}()},
\code{\link{label_1d_grid}()},
\code{\link{label_1d_loon}()},
\code{\link{lines_1d_graphics}()},
\code{\link{lines_1d_grid}()},
\code{\link{lines_1d_loon}()},
\code{\link{points_1d_graphics}()},
\code{\link{points_1d_grid}()},
\code{\link{points_1d_loon}()},
\code{\link{rect_1d_graphics}()},
\code{\link{rect_1d_grid}()},
\code{\link{rect_1d_loon}()},
\code{\link{rug_1d_graphics}()},
\code{\link{rug_1d_grid}()}
}
\author{
Marius Hofert and Wayne Oldford
}
\concept{default 1d plot functions}
\concept{default 1d plot functions using the interactive loon package}
