% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{turn_checker}
\alias{turn_checker}
\title{Check the Turns (Number/Type)}
\usage{
turn_checker(turns, n2dplots, first1d, last1d)
}
\arguments{
\item{turns}{The turns}

\item{n2dplots}{The number of 2d plots}

\item{first1d}{A logical indicating whether the first 1d plot should be plotted}

\item{last1d}{A logical indicating whether the last 1d plot should be plotted}
}
\value{
TRUE (unless it fails)
}
\description{
Check the Turns (Number/Type)
}
\seealso{
Other zenplot technical tools: 
\code{\link{convert_occupancy}()},
\code{\link{is.standard}()},
\code{\link{n2dcols_aux}()},
\code{\link{num_cols}()}
}
\author{
Marius Hofert
}
\concept{zenplot technical tools}
