% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{zenarrow}
\alias{zenarrow}
\title{Defining an arrow}
\usage{
zenarrow(turn, angle = 80, length = 1, coord.scale = 1)
}
\arguments{
\item{turn}{The direction in which the arrow will point ("l", "r", "d", "u")}

\item{angle}{The angle}

\item{length}{The length of the arrow in [0,1] from tip to base}

\item{coord.scale}{Scale the coordinates of the arrow}
}
\value{
A 3-column matrix containing the (x,y) coordinates of the left
        edge end point, the arrow head and the right edge end point
}
\description{
Defining an arrow
}
\seealso{
Other graphical tools: \code{\link{l_ispace_config}},
  \code{\link{na_omit_loon}}, \code{\link{plot_region}},
  \code{\link{vport}}
}
\author{
Marius Hofert
}
\concept{graphical tools}
