% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenplots.R
\docType{package}
\name{zenplots-package}
\alias{zenplots}
\alias{zenplots-package}
\title{zenplots: Zigzag Expanded Navigation Plots}
\description{
Zenplots, like pairs plots (scatterplot matrices), lay out a large
number of one- and two-dimensional plots in an organized way.
}
\details{
Unlike pairs plots, zenplots can lay out a much larger number of
plots by pursuing a zigzagging layout (following a zenpath) of
alternating one- and two-dimensional plots.

The plots can be created by R's base graphics, by the grid graphics
system, or made interactive (brushing, etc.) using the \pkg{loon} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://great-northern-diver.github.io/zenplots}
  \item \url{https://github.com/great-northern-diver/zenplots}
}

}
\author{
\strong{Maintainer}: Wayne Oldford \email{rwoldford@uwaterloo.ca}

Authors:
\itemize{
  \item Marius Hofert \email{marius.hofert@uwaterloo.ca}
}

}
\keyword{internal}
