% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenplots.R
\docType{package}
\name{zenplots}
\alias{zenplots}
\alias{zenplots-package}
\title{zenplots: Zigzag Expanded Navigation Plots}
\description{
Zenplots, like pairs plots (scatterplot matrices), lay out a large
number of one- and two-dimensional plots in an organized way.
}
\details{
Unlike pairs plots, zenplots can lay out a much larger number of
plots by pursuing a zigzagging layout (following a zenpath) of
alternating one- and two-dimensional plots.

The plots can be created by R's base graphics package, by the grid
graphics package, or even made interactive (brushing, etc.) by using
using the loon package.
}
