% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract_2d}
\alias{extract_2d}
\title{Extracting information for our default/provided plot2d()}
\usage{
extract_2d(zargs)
}
\arguments{
\item{zargs}{The argument list as passed from zenplot(); this must at least
contain 'x', 'vars', 'num', 'lim' and 'labs'.}
}
\value{
A list with
        - the data x and y to plot in the 2d plot;
        - a list with all columns of x;
        - the group numbers (for each column of x);
        - the variable numbers (for each column of x);
        - the group labels (for each column of x);
        - the variable labels (for each column of x);
        - the x- and y-axis limits;
        - a logical indicating whether x and y are in the same group
}
\description{
Extracting information for our default/provided plot2d()
}
\note{
Performance critical
}
\seealso{
Other tools for constructing your own plot1d and plot2d functions: \code{\link{burst_aux}},
  \code{\link{burst}}, \code{\link{check_zargs}},
  \code{\link{extract_1d}}, \code{\link{plot_indices}}
}
\author{
Marius Hofert
}
\concept{tools for constructing your own plot1d and plot2d functions}
