% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_indices}
\alias{plot_indices}
\title{Plot Indices of the Current Plot}
\usage{
plot_indices(zargs)
}
\arguments{
\item{zargs}{argument list as passed from \code{\link{zenplot}()}.
This must at least contain \code{vars} and \code{num}; see
\code{\link{zenplot}()} for an explanation of these variables..}
}
\value{
A \code{numeric(2)} containing the indices of the x and y variables to
be plotted in the current plot (the plot with number \code{num}). If
the current plot is a 2d plot, the same variable is used twice.
}
\description{
Determining the indices of the x and y variables of the current plot
}
\details{
This is an auxiliary function useful, for example, when writing
user-provided 1d and 2d plot functions.
}
\note{
This is exported so that one doesn't always have to figure
      out whether the variables (axes) in the current plot need
      to be switched manually.
}
\seealso{
Other tools for constructing your own plot1d and plot2d functions: \code{\link{burst_aux}},
  \code{\link{burst}}, \code{\link{check_zargs}},
  \code{\link{extract_1d}}, \code{\link{extract_2d}}
}
\author{
Marius Hofert
}
\concept{tools for constructing your own plot1d and plot2d functions}
\keyword{datagen}
